create useSetGlobalCommandMenuContext
